+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ## 5.1.0-beta.1
 
 * Feat: Measure app start time (#1487)
-* Feat: Automatic breadcrumbs logging for fragment lifecycle (#1522) 
+* Feat: Automatic breadcrumbs logging for fragment lifecycle (#1522)
+* Feat: Support transaction waiting for children to finish. (#1535) 
 
 ## 5.0.1